requirements_gui.txt
# Discord Bot GUI 依赖包
PyQt6>=6.5.0
requests>=2.28.0
python-dotenv>=1.0.0
pyinstaller>=5.10.0

# 可选的图标和资源处理
Pillow>=9.0.0

# 网络请求和JSON处理（通常是内置的，但确保兼容性）
urllib3>=1.26.0
certifi>=2022.0.0

# 日志和文件处理
pathlib2>=2.3.0